チャンネルをリアルタイムに取得（完成）
