+++ ui/src/provider/channel-provider.ts
@@ -1,52 +1,33 @@
 import { db } from "../lib/firebase";
-import {collection, query, orderBy, onSnapshot, getDocs } from 'firebase/firestore'
+import {collection,query, where,onSnapshot } from "firebase/firestore";
 import Channel from "../models/Channel"
 
-/**
- * Read: チャンネルのリストを取得する
- * @returns channelArr
- */
-export async function getChannelList() {
+//Get a list of cities from your database
+export function getChannelList(
+    uid: number[],
+		setChannel: React.Dispatch<React.SetStateAction<Channel[]>>
+	) {
+    const docsQuery = query(
+      collection(db, "channels"),
+      where("id","in",uid)
+    );
 
-	//最後にstate(Channel)に格納するためのChannel配列を作成
-	let channelArr: Channel[] = [];
+    onSnapshot(docsQuery, (querySnapshot) => {
 
-	/**
-	 * どのコレクションからどのような条件で取得するかを決める
-	 */
-		const q = query(collection(db, "channels"),orderBy('id','asc'));
+			const channels: Channel[] = [];
 
-	/**
-	 * getDocsはコレクションを取得（コレクションはDBで言うテーブルみたいなもの）
-	 * getDocはドキュメントを取得（ドキュメントはDBで言う行みたいなもの）
-	 */
-	const querySnapshot = await getDocs(q);
+      querySnapshot.forEach((doc) => {
+        const newChannel = doc.data() as Channel;
 
-	/**
-	 * 取得したコレクションの中にあるドキュメントをすべてループさせる
-	 * channelはドキュメント
-	 */
-	querySnapshot.forEach((channel) => {
-		// データが取得できているか確認するときに使用する。
-		//console.log(channel.id, " => ", channel.data());
-		//const newChannel = channel.data() as Channel;
-		const newChannel = {} as Channel;
+        newChannel.id = doc.data().id;
+        newChannel.name = doc.data().name;
+        newChannel.desc = doc.data().desc;
 
-		/**
-		 * doc.data()は、クエリドキュメントスナップショットに対して未定義になることはない
-		 *
-		 * つまり channel.data().desc など descではなく全く違う言葉(appleなど)にしてもコード上ではエラーになることはない
-		 */
-		newChannel.id = channel.data().id;
-		newChannel.name = channel.data().name;
-		newChannel.desc = channel.data().desc;
+				//console.log(newChannel);
 
+        channels.push(newChannel);
+      });
 
-		/**
-		 * セットした値を配列に格納する
-		 */
-		channelArr.push(newChannel);
-	});
-
-	return channelArr;
-}
+      setChannel(channels);
+    });
+  }
